+++ packages/pangea-node-sdk/CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `rotation_state` field support in key rotation in Vault service.
 - Vault service. Post quantum signing algorithms support
+- AuthZ service support
 
 ### Changed